+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to similar are documented here.
 ## 0.5.0
 
 * Add `DiffOp::apply_to_hook` to apply a captured op to a diff hook.
+* Added virtual newline handling to `iter_changes`.
 
 ## 0.4.0